backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
zxcvbn==4.4.28
fastapi-mail==1.4.1

# ML and Data Science
torch==2.1.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2

# DSPy (for LLM orchestration)
dspy-ai==2.3.5

# Ollama (local LLM)
requests==2.31.0  # For Ollama HTTP requests

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0
python-docx==1.1.0
PyPDF2==3.0.1
reportlab>=4.0.0  # For generating demo PDFs
slowapi==0.1.9
bleach==6.1.0
python-magic==0.4.27
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.40.0
celery==5.3.4
redis==5.0.1

# Load Testing
locust==2.20.1

# CORS (handled by FastAPI middleware)
